+++ cocoa-beans-state/src/main/java/net/apartium/cocoabeans/state/ListObservable.java
@@ -2,6 +2,7 @@ package net.apartium.cocoabeans.state;
 
 import org.jetbrains.annotations.ApiStatus;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -20,4 +21,12 @@ public interface ListObservable<E> extends CollectionObservable<E, List<E>>, Lis
 
     @Override
     <T> ListObservable<E> filter(Function<E, Observable<T>> mapper, Predicate<T> filter);
+
+    @ApiStatus.AvailableSince("0.0.50")
+    @Override
+    <R> ListObservable<R> mapEach(Function<E, R> mapper);
+
+    @ApiStatus.AvailableSince("0.0.50")
+    @Override
+    <R> ListObservable<R> flatMapEach(Function<E, Observable<R>> mapper);
 }