+++ cocoa-beans-state/src/main/java/net/apartium/cocoabeans/state/SetMapEachObservable.java
@@ -1,13 +1,14 @@
 package net.apartium.cocoabeans.state;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
-/* package-private */ class SetFilterObservable<E> extends FilterObservable<E, Set<E>> implements SetObservable<E> {
+/* package-private */ class SetMapEachObservable<F, E> extends MapElementObservable<F, E, Set<E>> implements SetObservable<E> {
 
-    public SetFilterObservable(Observable<Set<E>> base, Function<E, Observable<Boolean>> filter, Function<Collection<E>, Set<E>> copyOf, Function<Integer, ? extends Collection<E>> createInitSet) {
-        super(base, filter, copyOf, createInitSet);
+    public SetMapEachObservable(Observable<? extends Collection<F>> base, Function<F, E> mapper, Function<Collection<E>, Set<E>> collectionMapper, Function<Integer, ? extends Collection<E>> constructCollection) {
+        super(base, mapper, collectionMapper, constructCollection);
     }
 
     @Override