+++ cocoa-beans-state/src/main/java/net/apartium/cocoabeans/state/SetObservable.java
@@ -2,6 +2,7 @@ package net.apartium.cocoabeans.state;
 
 import org.jetbrains.annotations.ApiStatus;
 
+import java.util.Collection;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -21,4 +22,11 @@ public interface SetObservable<E> extends CollectionObservable<E, Set<E>> {
     @Override
     <T> SetObservable<E> filter(Function<E, Observable<T>> mapper, Predicate<T> filter);
 
+    @ApiStatus.AvailableSince("0.0.50")
+    @Override
+    <R> SetObservable<R> mapEach(Function<E, R> mapper);
+
+    @ApiStatus.AvailableSince("0.0.50")
+    @Override
+    <R> SetObservable<R> flatMapEach(Function<E, Observable<R>> mapper);
 }